% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_bnpmediation.R
\name{bnpm}
\alias{bnpm}
\title{Posterior Means and 95\% C.I.s of the NIE, NDE and TE}
\usage{
bnpm(dataTreatment, dataControl, prior, mcmc, state, status = TRUE, na.action,
  q = 2, NN = 10, n1 = 10, n0 = 10, extra.thin = 0)
}
\arguments{
\item{q}{A dimension of the observed data, i.e., number of covariates plus 2}

\item{NN}{Number of samples drawn for each iteration from the joint distribution of the mediator and the covariates. Default is 10.}

\item{n1}{Number of observations under Z=1}

\item{n0}{Number of observations under Z=0}

\item{extra.thin}{Giving the extra thinning interval}

\item{obj1}{The fitted model of the observed data under Z=1 from DPdensity}

\item{obj0}{The fitted model of the observed data under Z=0 from Dpdensity}
}
\value{
ENIE Posterior mean of the NIE

ENDE Posterior mean of the NDE

ETE Posterior mean of the TE

IE.c.i 95\% C.I. of the NIE

DE.c.i 95\% C.I. of the NDE

TE.c.i 95\% C.I. of the TE

Y11 Posterior samples of Y11

Y00 Posterior samples of Y00

Y10 Posterior samples of Y10
}
\description{
Obtain posterior means and credible intervals of the effects.
}
