% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_ploteffects.R
\name{plot_effects}
\alias{plot_effects}
\title{Posterior plots of the effects
To obtain the posterior plots of the effects}
\usage{
plot_effects(obj)
}
\arguments{
\item{obj}{The fitted model (For example from the output from bnpmediation or bnpconmediation functions)}
}
\value{
plots Draws the plot for the models
}
\description{
Posterior plots of the effects
To obtain the posterior plots of the effects
}
